feat: create BaseInfrastructureV2 to demonstrate before/after
